requirements.dev.txt
black
bokeh
fastparquet
isort
loguru
mypy
numpy
pandas
pdoc
pylint
pynacl
pytest
safetensors
scikit-learn
tensorflow
torch
types-setuptools